+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Phone number details to Resume Candidate info
 - "data" to job description create
+- Custom fields to Job Descriptions
 
 ## [5.3.1] - 2023-03-30
 ### Changed